+++ www/src/ui/splash/splash_mesh_1.ts
@@ -1,0 +1,5 @@
+import {Mesh} from "vauxcel";
+
+export class SplashShader1 extends Mesh {
+
+}
